# Upgrade Rust toolchain to `nightly-2024-03-01`