--- a/tests/expected/intrinsics/simd-rem-div-zero/main.rs
+++ b/tests/expected/intrinsics/simd-rem-div-zero/main.rs
@@ -2,16 +2,13 @@
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
 //! Checks that `simd_rem` triggers a failure when the divisor is zero
-#![feature(repr_simd, platform_intrinsics)]
+#![feature(repr_simd, core_intrinsics)]
+use std::intrinsics::simd::simd_rem;
 
 #[repr(simd)]
 #[allow(non_camel_case_types)]
 #[derive(Clone, Copy, PartialEq, Eq)]
 pub struct i32x2(i32, i32);
-
-extern "platform-intrinsic" {
-    fn simd_rem<T>(x: T, y: T) -> T;
-}
 
 #[kani::proof]
 fn test_simd_rem() {