--- a/tests/kani/Intrinsics/SIMD/Operators/arith.rs
+++ b/tests/kani/Intrinsics/SIMD/Operators/arith.rs
@@ -1,21 +1,15 @@
 // Copyright Kani Contributors
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
-//! This test doesn't work because support for SIMD intrinsics isn't available
-//! at the moment in Kani. Support to be added in
-//! <https://github.com/model-checking/kani/issues/1148>
-#![feature(repr_simd, platform_intrinsics)]
+//! Checks that the SIMD intrinsics `simd_add`, `simd_sub` and
+//! `simd_mul` are supported and return the expected results.
+#![feature(repr_simd, core_intrinsics)]
+use std::intrinsics::simd::{simd_add, simd_mul, simd_sub};
 
 #[repr(simd)]
 #[allow(non_camel_case_types)]
 #[derive(Clone, Copy, PartialEq, Eq)]
 pub struct i8x2(i8, i8);
-
-extern "platform-intrinsic" {
-    fn simd_add<T>(x: T, y: T) -> T;
-    fn simd_sub<T>(x: T, y: T) -> T;
-    fn simd_mul<T>(x: T, y: T) -> T;
-}
 
 macro_rules! verify_no_overflow {
     ($cf: ident, $uf: ident) => {{