Upgrades the Rust toolchain to `nightly-2024-03-01`. The Rust compiler PRs that triggered changes in this upgrades are:
 * https://github.com/rust-lang/rust/pull/121516
 * https://github.com/rust-lang/rust/pull/121598
 * https://github.com/rust-lang/rust/pull/121489
 * https://github.com/rust-lang/rust/pull/121783

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 and MIT licenses.