@adpaco-aws 2024-03-01
Unfortunately I'm getting consistent >400% regressions in `solver_runtime` for `s2n-quic/quic/s2n-quic-core/recovery::rtt_estimator::test::weighted_average_test` (see [this CI run](https://github.com/model-checking/kani/actions/runs/8115665074) for example). Other values like program steps, VCCs, etc. haven't really changed.