@zhassan-aws 2024-03-02
This test is known to have high variance in performance, e.g. see https://github.com/model-checking/kani/actions/runs/7453782015

The numbers in this PR are on par with the numbers in previous PRs, so I would consider this to be noise. You can confirm by running the test locally a number of times with and without the changes.