@adpaco-aws 2024-03-04
Thanks all for your feedback. I've run each version 10 times and the results are:
| iter. | before | after |
| ---   | ---    | ---   |
| 1 | 150.85266 | 200.95573 |
| 2 | 150.59746 | 200.90263 |
| 3 | 150.4481 | 202.78355 |
| 4 | 150.52237 | 201.8042 |
| 5 | 150.50912 | 200.75551 |
| 6 | 150.19904 | 200.7821 |
| 7 | 151.47997 | 200.9002 |
| 8 | 150.85907 | 200.75055 |
| 9 | 150.36966 | 204.47491 |
| 10 | 150.9473 | 201.89137 |

So it looks like there's a performance regression of about +50s/+33% for this test.