@adpaco-aws 2024-03-04
After some discussion, we consider that **the regression is due to SAT variability**.

No data suggests that the problem is more difficult per se after the toolchain upgrade. For example, runtime symex remains constant in all cases (~0.8s), as does program size (16292 steps) and VCCs (615/231).

The table below shows SAT-solving time for the UNSATISFIABLE instance in each of the iterations.

| iter. | before | after |
| --- | --- | --- |
| 1 | 146.51s | 196.37s |
| 2 | 146.25s | 196.281s |
| 3 | 146.032s | 198.212s |
| 4 | 146.176s | 197.226s |
| 5 | 146.168s | 196.173s |
| 6 | 145.878s | 196.158s |
| 7 | 147.131s | 196.314s |
| 8 | 146.485s | 196.165s |
| 9 | 146.028s | 199.885s |
| 10 | 146.612s | 197.307s |

If we compare this with the table posted in my previous comment, we can conclude that the total time increase is due to the increase in SAT-solving for the UNSATISFIABLE instance. Therefore, there is no reason to block this toolchain upgrade.

@zhassan-aws pointed out that there's almost no difference before/after this toolchain upgrade if we use CaDiCaL instead (the current solver is Kissat). So we'll propose switching the solver to CaDiCaL for this test.